Added Pointer Json file & new exe
Added pointer Json file so its easier to update the pointer in case it doesnt work in the future check here often to see new updates. New exe so it can read from the json file. Be sure to have both in the same directory it should automatically create a file called totalkills.txt